+++ src/core.rs
@@ -1,12 +1,12 @@
 //! Provides the components and functions required to work with memory mapped data channels.
-mod header;
+mod metadata;
 mod reader;
 mod tick;
 mod utils;
 mod version;
 mod writer;
 
-pub use header::*;
+pub use metadata::*;
 pub use reader::*;
 pub use tick::*;
 pub use writer::*;
@@ -40,16 +40,15 @@ use std::result::Result;
 ///
 /// ```
 /// # use kekbit::core::TickUnit::Nanos;
-/// # use kekbit::core::Header;
 /// use kekbit::core::*;
 /// # const FOREVER: u64 = 99_999_999_999;
 /// let writer_id = 1850;
 /// let channel_id = 42;
-/// # let header = Header::new(writer_id, channel_id, 300_000, 1000, FOREVER, Nanos);
+/// # let metadata = Metadata::new(writer_id, channel_id, 300_000, 1000, FOREVER, Nanos);
 /// let test_tmp_dir = tempdir::TempDir::new("kektest").unwrap();
-/// # let writer = shm_writer(&test_tmp_dir.path(), &header).unwrap();
+/// # let writer = shm_writer(&test_tmp_dir.path(), &metadata).unwrap();
 /// let reader = shm_reader(&test_tmp_dir.path(), channel_id).unwrap();
-/// println!("{:?}", reader.header());
+/// println!("{:?}", reader.metadata());
 ///
 /// ```
 pub fn shm_reader(root_path: &Path, channel_id: u64) -> Result<ShmReader, ChannelError> {
@@ -106,18 +105,17 @@ pub fn shm_reader(root_path: &Path, channel_id: u64) -> Result<ShmReader, Channe
 ///
 /// ```
 /// # use kekbit::core::TickUnit::Nanos;
-/// # use kekbit::core::Header;
 /// use kekbit::core::*;
 /// # const FOREVER: u64 = 99_999_999_999;
 /// let writer_id = 1850;
 /// let channel_id = 42;
-/// # let header = Header::new(writer_id, channel_id, 300_000, 1000, FOREVER, Nanos);
+/// # let metadata = Metadata::new(writer_id, channel_id, 300_000, 1000, FOREVER, Nanos);
 /// let test_tmp_dir = tempdir::TempDir::new("kektest").unwrap();
-/// # let writer = shm_writer(&test_tmp_dir.path(), &header).unwrap();
+/// # let writer = shm_writer(&test_tmp_dir.path(), &metadata).unwrap();
 /// let duration = 1000;
 /// let tries = 10;
 /// let reader = try_shm_reader(&test_tmp_dir.path(), channel_id, duration, tries).unwrap();
-/// println!("{:?}", reader.header());
+/// println!("{:?}", reader.metadata());
 ///
 /// ```
 pub fn try_shm_reader(root_path: &Path, channel_id: u64, duration_millis: u64, tries: u64) -> Result<ShmReader, ChannelError> {
@@ -141,7 +139,7 @@ pub fn try_shm_reader(root_path: &Path, channel_id: u64, duration_millis: u64, t
 /// # Arguments
 ///
 /// * `root_path` - The path to the folder where all the channels will be stored grouped by writers id.
-/// * `header` - a structure of type [Header](struct.Header.html) which contains the complete information required to create a channel.
+/// * `metadata` - a structure of type [Metadata](struct.Metadata.html) which contains the complete information required to create a channel.
 ///
 /// # Errors
 ///
@@ -152,7 +150,6 @@ pub fn try_shm_reader(root_path: &Path, channel_id: u64, duration_millis: u64, t
 /// ```
 /// use kekbit::core::TickUnit::Nanos;
 /// use kekbit::core::*;
-/// use kekbit::core::Header;
 /// use kekbit::api::Writer;
 ///
 /// const FOREVER: u64 = 99_999_999_999;
@@ -160,13 +157,13 @@ pub fn try_shm_reader(root_path: &Path, channel_id: u64, duration_millis: u64, t
 /// let channel_id = 42;
 /// let capacity = 3000;
 /// let max_msg_len = 100;
-/// let header = Header::new(writer_id, channel_id, capacity, max_msg_len, FOREVER, Nanos);
+/// let metadata = Metadata::new(writer_id, channel_id, capacity, max_msg_len, FOREVER, Nanos);
 /// let test_tmp_dir = tempdir::TempDir::new("kektest").unwrap();
-/// let mut writer = shm_writer(&test_tmp_dir.path(), &header).unwrap();
+/// let mut writer = shm_writer(&test_tmp_dir.path(), &metadata).unwrap();
 /// writer.heartbeat().unwrap();
 /// ```
-pub fn shm_writer(root_path: &Path, header: &Header) -> Result<ShmWriter, ChannelError> {
-    let kek_file_path = storage_path(root_path, header.channel_id()).into_path_buf();
+pub fn shm_writer(root_path: &Path, metadata: &Metadata) -> Result<ShmWriter, ChannelError> {
+    let kek_file_path = storage_path(root_path, metadata.channel_id()).into_path_buf();
     if kek_file_path.exists() {
         return Err(StorageAlreadyExists {
             file_name: kek_file_path.to_str().unwrap().to_string(),
@@ -200,7 +197,7 @@ pub fn shm_writer(root_path: &Path, header: &Header) -> Result<ShmWriter, Channe
                 file_name: err.to_string(),
             })
         })?;
-    let total_len = (header.capacity() + header.len() as u32 + FOOTER_LEN) as u64;
+    let total_len = (metadata.capacity() + metadata.len() as u32 + FOOTER_LEN) as u64;
     kek_file.set_len(total_len).or_else(|err| {
         Err(CouldNotAccessStorage {
             file_name: err.to_string(),
@@ -210,7 +207,7 @@ pub fn shm_writer(root_path: &Path, header: &Header) -> Result<ShmWriter, Channe
     let mut mmap =
         unsafe { MmapOptions::new().map_mut(&kek_file) }.or_else(|err| Err(MemoryMappingFailed { reason: err.to_string() }))?;
     let buf = &mut mmap[..];
-    header.write_to(buf);
+    metadata.write_to(buf);
     mmap.flush().or_else(|err| Err(AccessError { reason: err.to_string() }))?;
     info!("Kekbit channel with store {:?} succesfully initialized", kek_file_path);
     let res = ShmWriter::new(mmap);
@@ -259,11 +256,11 @@ mod test {
 
     #[test]
     fn check_max_len() {
-        let header = Header::new(100, 1000, 300_000, 1000, FOREVER, Nanos);
+        let metadata = Metadata::new(100, 1000, 300_000, 1000, FOREVER, Nanos);
         let test_tmp_dir = TempDir::new("kektest").unwrap();
-        let writer = shm_writer(&test_tmp_dir.path(), &header).unwrap();
+        let writer = shm_writer(&test_tmp_dir.path(), &metadata).unwrap();
         let reader = shm_reader(&test_tmp_dir.path(), 1000).unwrap();
-        assert_eq!(writer.header(), reader.header());
+        assert_eq!(writer.metadata(), reader.metadata());
     }
 
     #[test]
@@ -271,9 +268,9 @@ mod test {
         INIT_LOG.call_once(|| {
             simple_logger::init().unwrap();
         });
-        let header = Header::new(100, 1000, 10000, 1000, FOREVER, Nanos);
+        let metadata = Metadata::new(100, 1000, 10000, 1000, FOREVER, Nanos);
         let test_tmp_dir = TempDir::new("kektest").unwrap();
-        let mut writer = shm_writer(&test_tmp_dir.path(), &header).unwrap();
+        let mut writer = shm_writer(&test_tmp_dir.path(), &metadata).unwrap();
         let txt = "There are 10 kinds of people: those who know binary and those who don't";
         let msgs = txt.split_whitespace();
         let mut msg_count = 0;
@@ -310,11 +307,11 @@ mod test {
         INIT_LOG.call_once(|| {
             simple_logger::init().unwrap();
         });
-        let header = Header::new(100, 1000, 10000, 1000, FOREVER, Nanos);
+        let metadata = Metadata::new(100, 1000, 10000, 1000, FOREVER, Nanos);
         let test_tmp_dir = TempDir::new("kektest").unwrap();
         let mut msg_count = 0;
         {
-            let mut writer = shm_writer(&test_tmp_dir.path(), &header).unwrap();
+            let mut writer = shm_writer(&test_tmp_dir.path(), &metadata).unwrap();
             let txt = "There are 10 kinds of people: those who know binary and those who don't";
             let msgs = txt.split_whitespace();
             for m in msgs {
@@ -398,8 +395,8 @@ mod test {
             let good_reader = try_shm_reader(&test_tmp_dir.path(), channel_id, 1000, 20);
             assert!(good_reader.is_err());
         });
-        let header = Header::new(100, 1000, 10000, 1000, FOREVER, Nanos);
-        shm_writer(&root_dir.path(), &header).unwrap();
+        let metadata = Metadata::new(100, 1000, 10000, 1000, FOREVER, Nanos);
+        shm_writer(&root_dir.path(), &metadata).unwrap();
         handle.join().unwrap();
     }
 }